Cleaned warnings Volume code

+++ src/Volume/HarmonicBasis.cpp
@@ -4,11 +4,10 @@
 #include "Spectra/MatOp/SparseSymMatProd.h"
 #include "Spectra/SymGEigsShiftSolver.h"
 #include "Eigenmodes.h"
- #include "pmp/Timer.h"
+#include "pmp/Timer.h"
 #include <igl/slice.h>
 
-double Franke(const Eigen::Vector3d& p)
-{
+double Franke(const Eigen::Vector3d &p) {
     double x = p(0), y = p(1), z = p(2);
 
     double cx2 = (9. * x - 2.) * (9. * x - 2.);
@@ -25,47 +24,45 @@ double Franke(const Eigen::Vector3d& p)
     double cz5 = (9. * z - 5.) * (9. * z - 5.);
 
     return (3. / 4.) *
-               exp(-(1. / 4.) * cx2 - (1. / 4.) * cy2 - (1. / 4.) * cz2) +
+           exp(-(1. / 4.) * cx2 - (1. / 4.) * cy2 - (1. / 4.) * cz2) +
            (3. / 4.) * exp(-(1. / 49.) * cx1 - (9. / 10.) * y - 1. / 10. -
                            (9. / 10.) * z - 1. / 10.) +
            (1. / 2.) *
-               exp(-(1. / 4.) * cx7 - (1. / 4.) * cy3 - (1. / 4.) * cz5) -
+           exp(-(1. / 4.) * cx7 - (1. / 4.) * cy3 - (1. / 4.) * cz5) -
            (1. / 5.) * exp(-cx4 - cy7 - cz5);
 }
 
-double FrankeLaplacian(Eigen::Vector3d& p)
-{
+double FrankeLaplacian(Eigen::Vector3d &p) {
     double x = p(0), y = p(1), z = p(2);
 
     return (243. * (-2299. + 1800. * x * (2. + 9. * x))) /
-               (480200. *
-                exp((9. * (12. + 10. * x * (2. + 9. * x) + 49. * y + 49. * z)) /
-                    490.)) -
+           (480200. *
+            exp((9. * (12. + 10. * x * (2. + 9. * x) + 49. * y + 49. * z)) /
+                490.)) -
            (486. *
             exp(-pow(4. - 9. * x, 2) - pow(7. - 9. * y, 2) -
                 pow(5. - 9. * z, 2)) *
             (59. + 6. * x * (-8. + 9. * x) + 6. * y * (-14. + 9. * y) +
              6. * z * (-10 + 9 * z))) /
-               5. +
+           5. +
            (81. *
             exp((-pow(7. - 9. * x, 2) - 9 * pow(1. - 3. * y, 2) -
                  pow(5. - 9. * z, 2)) /
                 4.) *
             (77. + 9. * x * (-14. + 9. * x) + 27. * y * (-2. + 3. * y) +
              9 * z * (-10. + 9. * z))) /
-               8. +
+           8. +
            (729. * (2. + 3. * x * (-4. + 9. * x) + 3. * y * (-4. + 9. * y) +
                     3. * z * (-4. + 9. * z))) /
-               (16. *
-                exp((3. * (4. + 3. * x * (-4. + 9. * x) +
-                           3. * y * (-4. + 9. * y) + 3. * z * (-4. + 9. * z))) /
-                    4.0));
+           (16. *
+            exp((3. * (4. + 3. * x * (-4. + 9. * x) +
+                       3. * y * (-4. + 9. * y) + 3. * z * (-4. + 9. * z))) /
+                4.0));
 }
 
-void setup_3D_harmonic_matrices(PolyhedralMesh& mesh,
-                                Eigen::SparseMatrix<double>& K,
-                                Eigen::SparseMatrix<double>& M)
-{
+void setup_3D_harmonic_matrices(PolyhedralMesh &mesh,
+                                Eigen::SparseMatrix<double> &K,
+                                Eigen::SparseMatrix<double> &M) {
     using namespace std;
 
     const int nc = mesh.numCells();
@@ -76,46 +73,46 @@ void setup_3D_harmonic_matrices(PolyhedralMesh& mesh,
 
     std::vector<HarmonicPolyhedron> harmonicPolyhedra(nc);
 
-    threadHelper<6>([&](const int i){
+    threadHelper<6>([&](const int i) {
         vector<Eigen::Vector3d> vertices;
         vector<vector<int>> faces;
         allCellVertices[i] = mesh.getCellGeometry(i, vertices, faces);
         harmonicPolyhedra[i] = HarmonicPolyhedron(vertices, faces);
 
-        if(i % 10 == 0) std::cout << i << " / " << nc << std::endl;
+        if (i % 10 == 0) std::cout << i << " / " << nc << std::endl;
 
     }, nc);
 
     // cannot parallelize quadrature initialization because tetgen seems to be not thread safe
-    for(int i = 0; i < nc; ++i) harmonicPolyhedra[i].initQuadrature();
+    for (int i = 0; i < nc; ++i) harmonicPolyhedra[i].initQuadrature();
 
-    threadHelper<6>([&](const int i){
+    threadHelper<6>([&](const int i) {
 
         harmonicPolyhedra[i].stiffnessMatrix(allK[i]);
         harmonicPolyhedra[i].massMatrix(allM[i]);
 
-        if(i % 10 == 0) std::cout << i << " / " << nc << std::endl;
+        if (i % 10 == 0) std::cout << i << " / " << nc << std::endl;
 
     }, nc);
 
     vector<Eigen::Triplet<double>> tripK, tripM;
 
 
-    for(int i = 0; i < mesh.numCells(); ++i) {
+    for (int i = 0; i < mesh.numCells(); ++i) {
 
-        auto& cellVertices = allCellVertices[i];
-        Eigen::MatrixXd& Ki = allK[i];
+        auto &cellVertices = allCellVertices[i];
+        Eigen::MatrixXd &Ki = allK[i];
 
-        for(int j = 0; j < cellVertices.size(); ++j) {
-            for(int k = 0; k < cellVertices.size(); ++k) {
+        for (int j = 0; j < (int) cellVertices.size(); ++j) {
+            for (int k = 0; k < (int) cellVertices.size(); ++k) {
                 tripK.emplace_back(cellVertices[j], cellVertices[k], Ki(j, k));
             }
         }
 
-        Eigen::MatrixXd& Mi = allM[i];
+        Eigen::MatrixXd &Mi = allM[i];
 
-        for(int j = 0; j < cellVertices.size(); ++j) {
-            for(int k = 0; k < cellVertices.size(); ++k) {
+        for (int j = 0; j < (int) cellVertices.size(); ++j) {
+            for (int k = 0; k < (int) cellVertices.size(); ++k) {
                 tripM.emplace_back(cellVertices[j], cellVertices[k], Mi(j, k));
             }
         }
@@ -128,25 +125,20 @@ void setup_3D_harmonic_matrices(PolyhedralMesh& mesh,
     M.setFromTriplets(tripM.begin(), tripM.end());
 }
 
-double solve_3D_Franke_harmonic(std::string& meshname)
-{
+double solve_3D_Franke_harmonic(std::string &meshname) {
     PolyhedralMesh mesh(meshname);
     Eigen::SparseMatrix<double> S, M;
     setup_3D_harmonic_matrices(mesh, S, M);
-//    M*=2.5;
-//    S*=-1.0;
     Eigen::VectorXd b(mesh.points.size());
 
-    for (int i = 0; i < mesh.points.size(); i++)
-    {
+    for (int i = 0; i < (int) mesh.points.size(); i++) {
         Eigen::Vector3d p(mesh.points[i][0], mesh.points[i][1],
                           mesh.points[i][2]);
         b(i) = -FrankeLaplacian(p);
     }
     b = M * b;
-    for (int i = 0; i < mesh.points.size(); i++)
-    {
-        if(mesh.isBoundaryVertex(i)){
+    for (int i = 0; i < (int) mesh.points.size(); i++) {
+        if (mesh.isBoundaryVertex(i)) {
             // right-hand side: fix boundary values with franke function of the vertices
             Eigen::Vector3d p(mesh.points[i][0], mesh.points[i][1],
                               mesh.points[i][2]);
@@ -156,22 +148,19 @@ double solve_3D_Franke_harmonic(std::string& meshname)
     // Adjust the right-hand-side to account for the locked nodes
     for (unsigned int i = 0; i < S.outerSize(); i++)
         for (Eigen::SparseMatrix<double>::InnerIterator iter(S, i); iter;
-             ++iter)
-        {
-            int row = iter.row();
-            int col = iter.col();
-            if (!mesh.isBoundaryVertex(row) && mesh.isBoundaryVertex(col))
-            {
+             ++iter) {
+            int row = (int) iter.row();
+            int col = (int) iter.col();
+            if (!mesh.isBoundaryVertex(row) && mesh.isBoundaryVertex(col)) {
                 b[iter.row()] -= b[iter.col()] * iter.value();
             }
         }
     // Adjust the system matrix to account for the locked nodes
     for (unsigned int i = 0; i < S.outerSize(); i++)
         for (Eigen::SparseMatrix<double>::InnerIterator iter(S, i); iter;
-             ++iter)
-        {
-            int row = iter.row();
-            int col = iter.col();
+             ++iter) {
+            int row = (int) iter.row();
+            int col = (int) iter.col();
             if (mesh.isBoundaryVertex(row)) iter.valueRef() = iter.row() == iter.col() ? 1. : 0.;
             else if (mesh.isBoundaryVertex(col)) iter.valueRef() = 0;
         }
@@ -184,8 +173,7 @@ double solve_3D_Franke_harmonic(std::string& meshname)
     Eigen::VectorXd x = solver.solve(b);
     std::cout << "Size x :" << x.size() << std::endl;
     double error = 0.0;
-    for (int i = 0; i < mesh.points.size(); i++)
-    {
+    for (int i = 0; i < (int) mesh.points.size(); i++) {
         Eigen::Vector3d p(mesh.points[i][0], mesh.points[i][1],
                           mesh.points[i][2]);
 //        std::cout << "x: " << x[i] << " Franke : " << Franke(p) << std::endl;
@@ -194,11 +182,11 @@ double solve_3D_Franke_harmonic(std::string& meshname)
 
     std::cout << "DoF " << mesh.points.size() << std::endl;
     std::cout << "Franke RMSE error inner vertices: "
-              << sqrt(error / (double)mesh.points.size()) << std::endl;
-    return sqrt(error / (double)mesh.points.size());
+              << sqrt(error / (double) mesh.points.size()) << std::endl;
+    return sqrt(error / (double) mesh.points.size());
 }
 
-double setup_3D_harmonic_stiffness_matrix(PolyhedralMesh &mesh, Eigen::SparseMatrix<double> &S, std::ofstream &file){
+double setup_3D_harmonic_stiffness_matrix(PolyhedralMesh &mesh, Eigen::SparseMatrix<double> &S, std::ofstream &file) {
     using namespace std;
     pmp::Timer t;
     double time = 0.0;
@@ -206,17 +194,17 @@ double setup_3D_harmonic_stiffness_matrix(PolyhedralMesh &mesh, Eigen::SparseMat
     const int nc = mesh.numCells();
 
     std::vector<Eigen::MatrixXd> allK(nc);
-     std::vector<std::vector<int>> allCellVertices(nc);
+    std::vector<std::vector<int>> allCellVertices(nc);
 
     std::vector<HarmonicPolyhedron> harmonicPolyhedra(nc);
     t.start();
-    threadHelper<8>([&](const int i){
+    threadHelper<8>([&](const int i) {
         vector<Eigen::Vector3d> vertices;
         vector<vector<int>> faces;
         allCellVertices[i] = mesh.getCellGeometry(i, vertices, faces);
         harmonicPolyhedra[i] = HarmonicPolyhedron(vertices, faces);
 
-        if(i % 10 == 0) std::cout << i << " / " << nc << std::endl;
+        if (i % 10 == 0) std::cout << i << " / " << nc << std::endl;
 
     }, nc);
     t.stop();
@@ -225,28 +213,27 @@ double setup_3D_harmonic_stiffness_matrix(PolyhedralMesh &mesh, Eigen::SparseMat
     t.start();
 //    std::cout << "Basis construction << t.elapsed() <<
     // cannot parallelize quadrature initialization because tetgen seems to be not thread safe
-    for(int i = 0; i < nc; ++i) harmonicPolyhedra[i].initQuadrature();
+    for (int i = 0; i < nc; ++i) harmonicPolyhedra[i].initQuadrature();
     t.stop();
     time += t.elapsed();
     file << t.elapsed() << ",";
     t.start();
-    threadHelper<8>([&](const int i){
+    threadHelper<8>([&](const int i) {
 
         harmonicPolyhedra[i].stiffnessMatrix(allK[i]);
-//        if(i % 10 == 0) std::cout << i << " / " << nc << std::endl;
 
     }, nc);
 
-    vector<Eigen::Triplet<double>> tripK ;
+    vector<Eigen::Triplet<double>> tripK;
 
 
-    for(int i = 0; i < mesh.numCells(); ++i) {
+    for (int i = 0; i < mesh.numCells(); ++i) {
 
-        auto& cellVertices = allCellVertices[i];
-        Eigen::MatrixXd& Ki = allK[i];
+        auto &cellVertices = allCellVertices[i];
+        Eigen::MatrixXd &Ki = allK[i];
 
-        for(int j = 0; j < cellVertices.size(); ++j) {
-            for(int k = 0; k < cellVertices.size(); ++k) {
+        for (int j = 0; j < (int) cellVertices.size(); ++j) {
+            for (int k = 0; k < (int) cellVertices.size(); ++k) {
                 tripK.emplace_back(cellVertices[j], cellVertices[k], Ki(j, k));
             }
         }
@@ -264,13 +251,12 @@ double setup_3D_harmonic_stiffness_matrix(PolyhedralMesh &mesh, Eigen::SparseMat
     return time;
 }
 
-double solve_harmonic_eigenvalue_problem(std::string& meshname,Eigen::VectorXd &evalues, std::string& meshname_file)
-{
+double solve_harmonic_eigenvalue_problem(std::string &meshname, Eigen::VectorXd &evalues, std::string &meshname_file) {
 
     PolyhedralMesh mesh(meshname);
     Eigen::SparseMatrix<double> S, M;
     setup_3D_harmonic_matrices(mesh, S, M);
-    S*=-1.0;
+    S *= -1.0;
     std::string filename = "eigenmodes_Harmonic-" + meshname_file + ".csv";
     std::ofstream ev_file(filename);
     ev_file << "computed,analytic,offset" << std::endl;
@@ -279,9 +265,8 @@ double solve_harmonic_eigenvalue_problem(std::string& meshname,Eigen::VectorXd &
 
     // collect indices of inner vertices
     std::vector<int> indices;
-    for (int i = 0; i < mesh.points.size(); i++)
-    {
-        if(!mesh.isBoundaryVertex(i)){
+    for (int i = 0; i < (int) mesh.points.size(); i++) {
+        if (!mesh.isBoundaryVertex(i)) {
             indices.push_back(i);
         }
     }
@@ -289,8 +274,7 @@ double solve_harmonic_eigenvalue_problem(std::string& meshname,Eigen::VectorXd &
     Eigen::VectorXi in(indices.size());
     std::cout << "inner indices: " << indices.size() << std::endl;
     //Rewrite indices to Eigen::Vector
-    for (int i = 0; i < indices.size(); ++i)
-    {
+    for (int i = 0; i < (int) indices.size(); ++i) {
         in(i) = indices[i];
     }
 
@@ -304,7 +288,7 @@ double solve_harmonic_eigenvalue_problem(std::string& meshname,Eigen::VectorXd &
     int num_eval = 34;
     int converge_speed = 5 * num_eval;
     // S and M are sparse
-    using OpType =  Spectra::SymShiftInvert<double, Eigen::Sparse, Eigen::Sparse>;
+    using OpType = Spectra::SymShiftInvert<double, Eigen::Sparse, Eigen::Sparse>;
     using BOpType = Spectra::SparseSymMatProd<double>;
     OpType op(S_in_in, M_in_in);
     BOpType Bop(M_in_in);
@@ -312,20 +296,19 @@ double solve_harmonic_eigenvalue_problem(std::string& meshname,Eigen::VectorXd &
     // Construct generalized eigen solver object, seeking three generalized
     // eigenvalues that are closest to zero. This is equivalent to specifying
     // a shift sigma = 0.0 combined with the SortRule::LargestMagn selection rule
-    Spectra::SymGEigsShiftSolver<OpType, BOpType,Spectra::GEigsMode::ShiftInvert> geigs(op, Bop, num_eval, converge_speed, 0.0);
+    Spectra::SymGEigsShiftSolver<OpType, BOpType, Spectra::GEigsMode::ShiftInvert> geigs(op, Bop, num_eval,
+                                                                                         converge_speed, 0.0);
     geigs.init();
     geigs.compute(Spectra::SortRule::LargestMagn);
 
     Eigen::VectorXd evectors, analytic;
     // Retrieve results
-    if (geigs.info() == Spectra::CompInfo::Successful)
-    {
+    if (geigs.info() == Spectra::CompInfo::Successful) {
         evalues = geigs.eigenvalues();
     }
     analytic_eigenvalues_unitBall(analytic, num_eval);
     double error = 0.0;
-    for (int i = 0; i < evalues.size(); i++)
-    {
+    for (int i = 0; i < evalues.size(); i++) {
         ev_file << -evalues(i) << "," << analytic(i) << ","
                 << -evalues(i) - analytic(i) << std::endl;
         std::cout << "Computed evalue: " << -evalues(i)
@@ -333,7 +316,7 @@ double solve_harmonic_eigenvalue_problem(std::string& meshname,Eigen::VectorXd &
 
         error += pow(-evalues(i) - analytic(i), 2);
     }
-    error = sqrt(error / (double)evalues.size());
+    error = sqrt(error / (double) evalues.size());
     std::cout << "Root mean squared error: " << error << std::endl;
     return error;
 }
